requirements.txt
wheel
discord.py